+++ packages/dsv-charts/src/data-structure/index.ts
@@ -1,0 +1,1 @@
+export * from './array';
