+++ packages/dsv-charts/src/index.ts
@@ -1,1 +1,1 @@
-export * from './charts';
+export * from './old-charts';
