+++ packages/dsv-charts/src/old-charts/default-config.ts
