+++ packages/dsv-charts/src/old-charts/index.ts
@@ -1,0 +1,4 @@
+export * from './register';
+export * from './array';
+export * from './stack';
+export * from './queue';
