+++ packages/dsv-charts/src/old-charts/queue/queue.ts
@@ -1,25 +1,24 @@
-import { BorderType } from './../../typings/theme/index';
+import { ArrowType } from '../../typings/theme/index';
 import 'd3-transition';
-import { easeCubic } from 'd3-ease';
-import { scaleBand, ScaleBand } from 'd3-scale';
+import { scaleBand, scaleLinear, ScaleBand, ScaleLinear } from 'd3-scale';
 import { Selection } from 'd3-selection';
+import { max } from 'd3-array';
 import { BaseChart } from '../base';
 import { Cartesian2Layout } from '@dsv-charts/layouts/cartesian2';
 
 import {
   IConfig,
   Cartesian2LayoutType,
-  StackItemType,
-  StackDataType,
+  ArrayItemType,
+  ArrayDataType,
   ITheme,
   TextType,
-  TransitionType,
 } from '@dsv-charts/typings';
 
-export class StackChart extends BaseChart {
+export class QueueChart extends BaseChart {
   private _layout: Cartesian2Layout;
-  private _data: StackDataType;
-  private _yScale: ScaleBand<string>;
+  private _data: ArrayDataType;
+  private _xScale: ScaleBand<string>;
   private _rectGroup: Selection<SVGGElement, unknown, null, undefined>;
   private _textGroup: Selection<SVGGElement, unknown, null, undefined>;
   private _containerGroup: Selection<SVGGElement, unknown, null, undefined>;
@@ -40,17 +39,17 @@ export class StackChart extends BaseChart {
     this._textGroup = this._layout.addGroup();
     this._containerGroup = this._layout.addGroup();
 
-    this._data = super.getConfigByKey('data') as StackDataType;
+    this._data = super.getConfigByKey('data') as ArrayDataType;
   }
 
-  public setData(data: StackDataType, ignoreLifeCircle = false) {
+  public setData(data: ArrayDataType, ignoreLifeCircle = false) {
     super.setData(data, ignoreLifeCircle);
-    this._data = super.getConfigByKey('data') as StackDataType;
+    this._data = super.getConfigByKey('data') as ArrayDataType;
   }
 
   public setConfig(customConfig: IConfig): void {
     super.setConfig(customConfig);
-    this._data = super.getConfigByKey('data') as StackDataType;
+    this._data = super.getConfigByKey('data') as ArrayDataType;
   }
 
   public async renderAsync() {
@@ -71,65 +70,57 @@ export class StackChart extends BaseChart {
   private renderScale() {
     const innerRect = this._layout.getInnerRect();
 
-    this._yScale = scaleBand(this._data.map((d) => d.key).reverse(), [
-      innerRect.innerTop,
-      innerRect.innerBottom,
-    ]).paddingInner(0.1);
+    this._xScale = scaleBand(
+      this._data.map((d) => d.key),
+      [0, innerRect.innerWidth]
+    ).padding(0.01);
 
     return this;
   }
 
   private renderRectGroup() {
     const innerRect = this._layout.getInnerRect();
+    const rect = this._layout.getRect();
     const colorScheme = this.getThemeByKey('colorScheme');
-    const { duration } = this.getConfigByKey('transition') as TransitionType;
+    const { duration } = this.getConfigByKey('transition');
 
     this._rectGroup.call((g) => {
       g.selectAll('rect')
-        .data(this._data, (d: StackItemType) => d.key)
+        .data(this._data, (d: ArrayItemType) => d.key)
         .join(
           (enter) =>
             enter
               .append('rect')
               .attr('x', innerRect.innerRight)
-              .attr('y', 0)
-              .attr('width', innerRect.innerWidth)
-              .attr('height', 0)
-              .attr('fill', 'opacity')
+              .attr('width', this._xScale.bandwidth())
+              .attr('height', 50)
+              .attr('fill', colorScheme[0])
+              .attr('y', () => rect.center[1] - 50)
               .transition()
-              .ease(easeCubic)
               .duration(duration)
-              .attr('fill', colorScheme[0])
-              .attr('x', innerRect.innerLeft)
-              .attr('height', innerRect.innerTop)
-              .transition()
-              .attr('x', innerRect.innerLeft)
-              .attr('y', (d) => this._yScale(d.key))
-              .attr('fill', colorScheme[0])
-              .attr('height', this._yScale.bandwidth())
-              .attr('width', innerRect.innerWidth)
+              .attr('height', 50)
+              .attr('y', () => rect.center[1] - 50)
+              .attr('x', (d) => this._xScale(d.key) + innerRect.innerLeft)
               .selection(),
           (update) =>
             update
               .transition()
               .duration(duration)
-              .attr('x', innerRect.innerLeft)
-              .attr('y', (d) => this._yScale(d.key))
               .attr('fill', colorScheme[0])
-              .attr('height', this._yScale.bandwidth())
-              .attr('width', innerRect.innerWidth)
+              .attr('width', this._xScale.bandwidth())
+              .attr('height', 50)
+              .attr('x', (d) => this._xScale(d.key) + innerRect.innerLeft)
+              .attr('y', () => rect.center[1] - 50)
               .selection(),
           (exit) =>
             exit
+              .attr('opacity', 1)
               .transition()
-              .ease(easeCubic)
               .duration(duration)
-              .attr('fill', colorScheme[0])
-              .attr('x', innerRect.innerLeft)
-              .attr('y', 0)
-              .attr('height', innerRect.innerTop)
+              .attr('opacity', 0)
+              .attr('x', -rect.right)
+              .attr('y', rect.center[0] - 50)
               .transition()
-              .attr('x', 0)
               .remove()
         );
     });
@@ -139,54 +130,44 @@ export class StackChart extends BaseChart {
 
   private renderTextGroup() {
     const innerRect = this._layout.getInnerRect();
+    const rect = this._layout.getRect();
     const text = this.getThemeByKey('text') as TextType;
     const { duration } = this.getConfigByKey('transition');
 
     this._textGroup.call((g) => {
       g.selectAll('text')
-        .data(this._data, (d: StackItemType) => d.key)
+        .data(this._data, (d: ArrayItemType) => d.key)
         .join(
           (enter) =>
             enter
               .append('text')
-              .attr('y', 0)
-              .attr('x', () => innerRect.innerRight)
-              .attr('dx', () => innerRect.innerWidth / 2)
+              .attr('font-size', 20)
+              .attr('text-anchor', 'middle')
+              .attr('dx', this._xScale.bandwidth() / 2)
+              .attr('dy', 20)
+              .attr('x', rect.right)
+              .attr('y', () => rect.center[1])
               .transition()
-              .ease(easeCubic)
-              .duration(duration)
-              .attr('x', () => innerRect.innerLeft)
-              .attr('y', innerRect.innerTop / 2)
-              .transition()
-              .attr('dx', () => innerRect.innerWidth / 2)
-              .attr('y', (d) => this._yScale(d.key))
-              .attr('dy', (d) => this._yScale.bandwidth() / 2)
-              .attr('fill', text.textColor)
-              .selection()
-              .html((d) => d.name),
-          (update) =>
-            update
-              .transition()
-              .duration(duration)
-              .attr('x', () => innerRect.innerLeft)
-              .attr('dx', () => innerRect.innerWidth / 2)
-              .attr('y', (d) => this._yScale(d.key))
-              .attr('dy', (d) => this._yScale.bandwidth() / 2)
-              .attr('fill', text.textColor)
-              .selection()
-              .html((d) => d.name),
+              .attr('x', (d) => this._xScale(d.key) + innerRect.innerLeft)
+              .attr('y', () => rect.center[1]),
+          (update) => update,
           (exit) =>
             exit
               .transition()
-              .ease(easeCubic)
               .duration(duration)
-              .attr('x', innerRect.innerLeft)
-              .attr('y', 0)
-              .attr('dy', innerRect.innerTop / 2)
-              .transition()
-              .attr('x', 0)
+              .attr('x', -rect.right)
+              .attr('y', rect.center[1])
               .remove()
-        );
+        )
+        .transition()
+        .duration(duration)
+        .attr('x', (d) => this._xScale(d.key) + innerRect.innerLeft)
+        .attr('y', () => rect.center[1])
+        .attr('dx', this._xScale.bandwidth() / 2)
+        .attr('dy', 20)
+        .attr('fill', text.textColor)
+        .selection()
+        .html((d) => d.name);
     });
 
     return this;
@@ -194,15 +175,18 @@ export class StackChart extends BaseChart {
 
   private renderContainerGroup() {
     const innerRect = this._layout.getInnerRect();
-    const border = this.getThemeByKey('border') as BorderType;
+    const rect = this._layout.getRect();
+
+    const arrow = this.getThemeByKey('arrow') as ArrowType;
 
     const { duration } = this.getConfigByKey('transition');
 
     const points = [
-      [innerRect.innerLeft, innerRect.innerTop].join(' '),
-      [innerRect.innerLeft, innerRect.innerBottom].join(' '),
-      [innerRect.innerRight, innerRect.innerBottom].join(' '),
-      [innerRect.innerRight, innerRect.innerTop].join(' '),
+      [innerRect.innerLeft + 25, rect.center[1] - 85 - 25].join(' '),
+      [innerRect.innerLeft, rect.center[1] - 75 - 25].join(' '),
+      [innerRect.innerLeft + 25, rect.center[1] - 65 - 25].join(' '),
+      [innerRect.innerLeft, rect.center[1] - 75 - 25].join(' '),
+      [innerRect.innerRight, rect.center[1] - 75 - 25].join(' '),
     ].join(',');
 
     this._containerGroup
@@ -213,10 +197,8 @@ export class StackChart extends BaseChart {
       .duration(duration)
       .attr('points', (d) => d)
       .attr('fill', 'transparent')
-      .attr('stroke-linejoin', 'round')
-      .attr('stroke-width', border.width)
-      .attr('stroke-linecap', 'linecap')
-      .attr('stroke', border.color);
+      .attr('stroke-width', arrow.width)
+      .attr('stroke', arrow.color);
     return this;
   }
 
@@ -226,7 +208,6 @@ export class StackChart extends BaseChart {
     return this._rectGroup
       .transition()
       .duration(duration)
-      .transition()
       .on('end', () => {
         resolve();
       });
@@ -238,11 +219,10 @@ export class StackChart extends BaseChart {
     this._rectGroup.remove();
     this._containerGroup.remove();
 
-    this._yScale = null;
+    this._xScale = null;
     this._data = null;
     this._layout = null;
     this._textGroup = null;
     this._rectGroup = null;
-    this._containerGroup = null;
   }
 }
