--- packages/dsv-charts/src/re-charts/index.ts
@@ -1,5 +1,0 @@
-import { IArrayChartConfig } from './array';
-
-export type IConfig = IArrayChartConfig;
-
-export * from './array';
