--- packages/dsv-charts/src/re-types/index.ts
@@ -1,6 +1,0 @@
-export type ArrayLike<T> = {
-  [key: number]: T;
-  length: number;
-};
-
-export * from './IChart';
