+++ packages/dsv-website/src/config/data-structure/array/fill.js
@@ -1,13 +1,13 @@
 const code = `const array = dsv.create({
-  type: 'arrayChart',
-  data: [10, 8, 6, 4, 2, 1, 3, 5, 7, 9]
+type: 'array',
+data: [10, 8, 6, 4, 2, 1, 3, 5, 7, 9]
 })
-console.log(array)
-// array.fill(5, 5)
 
-// array.fill(15, 0, 5)
+array.fill(5, 5)
 
-// array.fill(30)
+array.fill(15, 0, 5)
+
+array.fill(30)
 `;
 
 module.exports = {
